requirements.txt
pocketflow>=0.0.1
openai>=1.0.0
scikit-learn>=1.0.0
pandas>=1.3.0
numpy>=1.20.0
streamlit>=1.20.0
faiss-cpu>=1.7.0
pyyaml>=6.0.0
anthropic>=0.23.2
google-cloud-aiplatform
google-genai